Apply styling and positioning for middle
Apply styling and positioning for middle section for
aesthetic purposes. Address vertical alignment in order
for elements in top section and middle section of the
right side of the display for consistency.Additionally,
modify changes in form's labels, width, and height for
better user experience.

Minor: Change background image for middle section to
better blend in with the other colors.